Move interface-types.ts to /common
