+++ extensions/ql-vscode/src/query-history/store/query-history-local-query-dto-mapper.ts
@@ -19,7 +19,7 @@ import {
   RawResultsSortState,
   SortDirection,
   SortedResultSetInfo,
-} from "../../pure/interface-types";
+} from "../../common/interface-types";
 
 export function mapLocalQueryItemToDomainModel(
   localQuery: QueryHistoryLocalQueryDto,
