+++ extensions/ql-vscode/src/variant-analysis/bqrs-processing.ts
@@ -3,7 +3,7 @@ import { Logger } from "../common/logging";
 import { transformBqrsResultSet } from "../common/bqrs-cli-types";
 import { AnalysisRawResults } from "./shared/analysis-result";
 import { MAX_RAW_RESULTS } from "./shared/result-limits";
-import { SELECT_TABLE_NAME } from "../pure/interface-types";
+import { SELECT_TABLE_NAME } from "../common/interface-types";
 
 export async function extractRawResults(
   cliServer: CodeQLCliServer,
