+++ extensions/ql-vscode/src/variant-analysis/run-remote-query.ts
@@ -18,7 +18,7 @@ import {
   UserCancellationException,
 } from "../common/vscode/progress";
 import { RequestError } from "@octokit/types/dist-types";
-import { QueryMetadata } from "../pure/interface-types";
+import { QueryMetadata } from "../common/interface-types";
 import { getErrorMessage, REPO_REGEX } from "../common/helpers-pure";
 import { getRepositoryFromNwo } from "./gh-api/gh-api-client";
 import {
