+++ extensions/ql-vscode/src/view/data-flow-paths/DataFlowPathsView.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import { useEffect, useState } from "react";
-import { ToDataFlowPathsMessage } from "../../pure/interface-types";
+import { ToDataFlowPathsMessage } from "../../common/interface-types";
 import { DataFlowPaths as DataFlowPathsDomainModel } from "../../variant-analysis/shared/data-flow-paths";
 import { DataFlowPaths } from "./DataFlowPaths";
