+++ extensions/ql-vscode/test/vscode-tests/no-workspace/local-queries/webview.test.ts
@@ -2,7 +2,7 @@ import { Uri, ViewColumn, WebviewPanel, window } from "vscode";
 import { basename } from "path";
 import { FileResult, fileSync } from "tmp";
 import { fileUriToWebviewUri } from "../../../../src/local-queries/webview";
-import { getDefaultResultSetName } from "../../../../src/pure/interface-types";
+import { getDefaultResultSetName } from "../../../../src/common/interface-types";
 
 describe("interface-utils", () => {
   describe("webview uri conversion", () => {
