+++ extensions/ql-vscode/test/vscode-tests/no-workspace/query-results.test.ts
@@ -17,7 +17,7 @@ import {
   DatabaseInfo,
   SortDirection,
   SortedResultSetInfo,
-} from "../../../src/pure/interface-types";
+} from "../../../src/common/interface-types";
 import { CodeQLCliServer, SourceInfo } from "../../../src/codeql-cli/cli";
 import { CancellationTokenSource, Uri } from "vscode";
 import { tmpDir } from "../../../src/tmp-dir";
